+++ src/device/tsp/Tsp/Script/SessionBaseExtensions/Methods_save.cs
@@ -107,7 +107,10 @@ public static partial class SessionBaseExtensionMethods
         }
     }
 
-    /// <summary>   A <see cref="Pith.SessionBase"/> extension method that saves a script to non-volatile-memory (NVM). </summary>
+    /// <summary>
+    /// A <see cref="Pith.SessionBase"/> extension method that saves a script to non-volatile-memory
+    /// (NVM).
+    /// </summary>
     /// <remarks>   2025-04-10. </remarks>
     /// <exception cref="ArgumentNullException">        Thrown when one or more required arguments
     ///                                                 are null. </exception>
@@ -115,7 +118,8 @@ public static partial class SessionBaseExtensionMethods
     ///                                                 invalid. </exception>
     /// <param name="session">      The session to act on. </param>
     /// <param name="scriptName">   Name of the script. </param>
-    public static void SaveScript( this SessionBase session, string scriptName )
+    /// <param name="skipSaved">    (Optional) [false] True to skip if the script is already saved. </param>
+    public static void SaveScript( this SessionBase session, string scriptName, bool skipSaved = false )
     {
         if ( session == null )
             throw new ArgumentNullException( nameof( session ) );
@@ -124,15 +128,23 @@ public static partial class SessionBaseExtensionMethods
         if ( string.IsNullOrWhiteSpace( scriptName ) )
             throw new ArgumentNullException( nameof( scriptName ) );
 
-        session.SetLastAction( $"checking if the {scriptName} script exists;. " );
+        session.SetLastAction( $"checking if the {scriptName} script was loaded;. " );
         session.LastNodeNumber = default;
         if ( session.IsNil( scriptName ) )
-            throw new InvalidOperationException( $"The script {scriptName} cannot be run because it was not found." );
+            throw new InvalidOperationException( $"The script {scriptName} cannot be saved because it was not loaded." );
 
         session.SetLastAction( $"checking if the {scriptName} saved script exists;. " );
         session.LastNodeNumber = default;
         if ( session.IsSavedScript( scriptName ) )
-            throw new InvalidOperationException( $"The script {scriptName} is already saved." );
+        {
+            if ( skipSaved )
+            {
+                session.TraceLastAction( $"Skipping {scriptName} script, which is already saved;. " );
+                return;
+            }
+            else
+                throw new InvalidOperationException( $"The script {scriptName} is already saved." );
+        }
 
         session.TraceLastAction( $"saving the {scriptName} script;. " );
         _ = session.WriteLine( $"{scriptName}.save()" );